requirements.txt
absl-py==0.13.0
aiofiles==0.7.0
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
astunparse==1.6.3
billiard==3.6.4.0
black==21.7b0
cached-property==1.5.2
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
et-xmlfile==1.1.0
fastapi==0.68.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.39.0
gunicorn==20.1.0
h11==0.12.0
h5py==2.10.0
httptools==0.2.0
idna==3.2
importlib-metadata==4.6.3
joblib==1.0.1
Keras-Preprocessing==1.1.2
kombu==5.1.0
Markdown==3.3.4
mypy-extensions==0.4.3
numpy==1.18.5
oauthlib==3.1.1
openpyxl==3.0.7
opt-einsum==3.3.0
pandas==1.3.1
pathspec==0.9.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.4.1
six==1.16.0
sklearn==0.0
starlette==0.14.2
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.2.0
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.3
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
